server/requirements.txt
flask~=2.1.3
python-dotenv~=0.21.0
aiohttp~=3.8.1
requests~=2.27.1